implemented field_name_attr and field_type_name_attr consistently along with _explicit_type_name attribute to ModelBase so that intermediate trash names don't pollute output
