+++ spyne/protocol/html/table/column.py
@@ -26,53 +26,10 @@ from inspect import isgenerator
 
 from lxml.html.builder import E
 
-from spyne import ModelBase, ByteArray, ComplexModelBase, Array, AnyUri, \
-    ImageUri
-from spyne.protocol.html import HtmlBase
+from spyne import ModelBase, ComplexModelBase, Array
 from spyne.util import coroutine, Break, urlencode
 from spyne.util.oset import oset
-from spyne.util.cdict import cdict
-
-
-class HtmlTableBase(HtmlBase):
-    def __init__(self, app=None, ignore_uncap=False, ignore_wrappers=True,
-            cloth=None, cloth_parser=None, header=True, table_name_attr='class',
-                     table_name=None, table_class=None, field_name_attr='class',
-              border=0, row_class=None, cell_class=None, header_cell_class=None,
-                  polymorphic=True, hier_delim='.', doctype=None, link_gen=None,
-                                         mrpc_delim_text='|', table_width=None):
-
-        super(HtmlTableBase, self).__init__(app=app,
-                     ignore_uncap=ignore_uncap, ignore_wrappers=ignore_wrappers,
-                cloth=cloth, cloth_parser=cloth_parser, polymorphic=polymorphic,
-                                         hier_delim=hier_delim, doctype=doctype)
-
-        self.header = header
-        self.table_name_attr = table_name_attr
-        self.table_name = table_name
-        self.field_name_attr = field_name_attr
-        self.border = border
-        self.row_class = row_class
-        self.cell_class = cell_class
-        self.header_cell_class = header_cell_class
-        self.link_gen = link_gen
-        self.table_class = table_class
-        self.table_width = table_width
-        self.mrpc_delim_text = mrpc_delim_text
-
-        if self.cell_class is not None and field_name_attr == 'class':
-            raise Exception("Either 'cell_class' should be None or "
-                            "field_name_attr should be != 'class'")
-
-        if self.header_cell_class is not None and field_name_attr == 'class':
-            raise Exception("Either 'header_cell_class' should be None or "
-                            "field_name_attr should be != 'class'")
-
-    def model_base_to_parent(self, ctx, cls, inst, parent, name,  **kwargs):
-        parent.write(self.to_unicode(cls, inst))
-
-    def null_to_parent(self, ctx, cls, inst, parent, name, **kwargs):
-        pass
+from spyne.protocol.html.table import HtmlTableBase
 
 
 class HtmlColumnTable(HtmlTableBase):
@@ -118,7 +75,6 @@ class HtmlColumnTable(HtmlTableBase):
         if before_table is not None:
             self.event_manager.add_listener("before_table", before_table)
 
-
     def model_base_to_parent(self, ctx, cls, inst, parent, name,
                                                       from_arr=False, **kwargs):
         inst_str = ''
@@ -127,14 +83,10 @@ class HtmlColumnTable(HtmlTableBase):
 
         if from_arr:
             td_attrs = {}
-            #if self.field_name_attr:
-            #    td_attrs[self.field_name_attr] = name
-            parent.write(E.tr(
-                E.td(
-                    inst_str,
-                    **td_attrs
-                )
-            ))
+
+            self.add_field_attrs(td_attrs, name, cls)
+
+            parent.write(E.tr(E.td(inst_str, **td_attrs)))
 
         else:
             parent.write(inst_str)
@@ -184,10 +136,11 @@ class HtmlColumnTable(HtmlTableBase):
                                                                sub_name, v, cls)
 
                 td_attrs = {}
-                if self.field_name_attr is not None:
-                    td_attrs[self.field_name_attr] = cls_attr.sub_name or k
+
+                self.add_field_attrs(td_attrs, cls_attr.sub_name or k, v)
+
                 if cls_attr.hidden:
-                    td_attrs['style'] = 'display:None'
+                    self.add_style(td_attrs, 'display:None')
 
                 with parent.element('td', td_attrs):
                     ret = self.to_parent(ctx, v, sub_value, parent,
@@ -249,15 +202,15 @@ class HtmlColumnTable(HtmlTableBase):
                 if issubclass(cls, ComplexModelBase):
                     fti = self.sort_fields(cls)
                     for k, v in fti:
-                        attr = self.get_cls_attrs(v)
-                        if attr.exc:
+                        cls_attr = self.get_cls_attrs(v)
+                        if cls_attr.exc:
                             continue
 
                         th_attrs = {}
-                        if self.field_name_attr is not None:
-                            th_attrs[self.field_name_attr] = k
-                        if attr.hidden:
-                            th_attrs['style'] = 'display:None'
+                        self.add_field_attrs(th_attrs, k, cls)
+
+                        if cls_attr.hidden:
+                            self.add_style(th_attrs, 'display:None')
 
                         header_name = self.trc(v, ctx.locale, k)
                         parent.write(E.th(header_name, **th_attrs))
@@ -268,9 +221,10 @@ class HtmlColumnTable(HtmlTableBase):
 
                 else:
                     th_attrs = {}
-                    if self.field_name_attr is not None:
-                        th_attrs[self.field_name_attr] = name
+                    self.add_field_attrs(th_attrs, name, cls)
+
                     header_name = self.trc(cls, ctx.locale, name)
+
                     parent.write(E.th(header_name, **th_attrs))
 
                 self.extend_header_row(ctx, cls, parent, name)
@@ -345,185 +299,3 @@ class HtmlColumnTable(HtmlTableBase):
 
     def extend_header_row(self, ctx, cls, parent, name, **kwargs):
         pass
-
-
-class HtmlRowTable(HtmlTableBase):
-    """Protocol that returns the response object as a html table.
-
-    The simple flavour is like the HtmlMicroFormatprotocol, but returns data
-    as a html table using the <table> tag.
-
-    Returns one record per table in a table with two columns.
-
-    :param app: A spyne.application.Application instance.
-    :param header: Boolean value to determine whether to show field
-        names in the beginning of the table or not. Defaults to True. Set to
-        False to skip headers.
-    :param table_name_attr: The name of the attribute that will contain the
-        response name of the complex object in the table tag. Set to None to
-        disable.
-    :param table_name: When not none, overrides what goes in `table_name_attr`.
-    :param table_class: When not none, specifies what goes in `class` attribute
-        in the `<table>` tag. Table name gets appended when
-        `table_name_attr == 'class'`
-    :param field_name_attr: The name of the attribute that will contain the
-        field names of the complex object children for every table cell. Set
-        to None to disable.
-    :param row_class: value that goes inside the <tr class="">
-    :param cell_class: value that goes inside the <td class="">
-    :param header_cell_class: value that goes inside the <th class="">
-    """
-
-    def __init__(self, *args, **kwargs):
-        super(HtmlRowTable, self).__init__(*args, **kwargs)
-
-        self.serialization_handlers = cdict({
-            ModelBase: self.model_base_to_parent,
-            AnyUri: self.anyuri_to_parent,
-            ImageUri: self.imageuri_to_parent,
-            ByteArray: self.not_supported,
-            ComplexModelBase: self.complex_model_to_parent,
-            Array: self.array_to_parent,
-        })
-
-    def model_base_to_parent(self, ctx, cls, inst, parent, name, from_arr=False,
-                                                                      **kwargs):
-        if from_arr:
-            td_attrib = {}
-            if False and self.field_name_attr:
-                td_attrib[self.field_name_attr] = name
-
-            parent.write(E.tr(E.td(self.to_unicode(cls, inst), **td_attrib)))
-        else:
-            parent.write(self.to_unicode(cls, inst))
-
-    @coroutine
-    def complex_model_to_parent(self, ctx, cls, inst, parent, name,
-                                                      from_arr=False, **kwargs):
-        attrib = {}
-        if self.table_name_attr is not None:
-            attrib[self.table_name_attr] = cls.get_type_name()
-        if self.table_width is not None:
-            attrib['width'] = self.table_width
-
-        with parent.element('table', attrib):
-            with parent.element('tbody'):
-                for k, v in self.sort_fields(cls):
-                    sub_attrs = self.get_cls_attrs(v)
-                    if sub_attrs.exc:
-                        logger.debug("\tExclude table cell %r type %r for %r",
-                                                                      k, v, cls)
-                        continue
-                    try:
-                        sub_value = getattr(inst, k, None)
-                    except:  # e.g. SQLAlchemy could throw NoSuchColumnError
-                        sub_value = None
-
-                    sub_name = v.Attributes.sub_name
-                    if sub_name is None:
-                        sub_name = k
-
-                    if sub_value is None and cls.Attributes.min_occurs == 0:
-                        self.null_to_parent(ctx, cls, sub_value, parent,
-                                                             sub_name, **kwargs)
-                        continue
-
-                    tr_attrib = {}
-                    if self.row_class is not None:
-                        tr_attrib['class'] = self.row_class
-
-                    with parent.element('tr', tr_attrib):
-                        th_attrib = {}
-                        if self.header_cell_class is not None:
-                            th_attrib['class'] = self.header_cell_class
-
-                        if self.field_name_attr is not None:
-                            th_attrib[self.field_name_attr] = sub_name
-
-                        if sub_attrs.hidden:
-                            th_attrib['style'] = 'display:None'
-
-                        if self.header:
-                            parent.write(E.th(
-                                self.trc(v, ctx.locale, sub_name),
-                                **th_attrib
-                            ))
-
-                        td_attrib = {}
-                        if self.cell_class is not None:
-                            td_attrib['class'] = self.cell_class
-
-                        if self.field_name_attr is not None:
-                            td_attrib[self.field_name_attr] = sub_name
-
-                        if sub_attrs.hidden:
-                            td_attrib['style'] = 'display:None'
-
-                        with parent.element('td', td_attrib):
-                            ret = self.to_parent(ctx, v, sub_value, parent,
-                                                            sub_name, **kwargs)
-                            if isgenerator(ret):
-                                try:
-                                    while True:
-                                        sv2 = (yield)
-                                        ret.send(sv2)
-                                except Break as b:
-                                    try:
-                                        ret.throw(b)
-                                    except StopIteration:
-                                        pass
-
-    @coroutine
-    def array_to_parent(self, ctx, cls, inst, parent, name, **kwargs):
-        with parent.element('div'):
-            if issubclass(cls, ComplexModelBase):
-                ret = super(HtmlRowTable, self).array_to_parent(
-                                         ctx, cls, inst, parent, name, **kwargs)
-                if isgenerator(ret):
-                    try:
-                        while True:
-                            sv2 = (yield)
-                            ret.send(sv2)
-                    except Break as b:
-                        try:
-                            ret.throw(b)
-                        except StopIteration:
-                            pass
-            else:
-                table_attrib = {}
-                if self.table_name_attr:
-                    table_attrib = {self.table_name_attr: name}
-                if self.table_width is not None:
-                    table_attrib['width'] = self.table_width
-
-                with parent.element('table', table_attrib):
-                    tr_attrib = {}
-                    if self.row_class is not None:
-                        tr_attrib['class'] = self.row_class
-                    with parent.element('tr', tr_attrib):
-                        if self.header:
-                            parent.write(E.th(self.trc(cls, ctx.locale,
-                                                          cls.get_type_name())))
-                        td_attrib = {}
-                        if self.cell_class is not None:
-                            td_attrib['class'] = self.cell_class
-                        cls_attrs = self.get_cls_attrs(cls)
-
-                        if cls_attrs.hidden:
-                            td_attrib['style'] = 'display:None'
-
-                        with parent.element('td', td_attrib):
-                            with parent.element('table'):
-                                ret = super(HtmlRowTable, self) \
-                                    .array_to_parent(ctx, cls, inst, parent,
-                                                                 name, **kwargs)
-                                if isgenerator(ret):
-                                    try:
-                                        while True:
-                                            sv2 = (yield)
-                                            ret.send(sv2)
-                                    except Break as b:
-                                        try:
-                                            ret.throw(b)
-                                        except StopIteration:
-                                            pass
